third_party/WebKit/LayoutTests/inspector/sources/debugger/script-failed-to-parse-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token }
Tests that FailedToParseScriptSource event is raised after compile script with syntax error.


Running: testScriptParsedEvent
Event with script-failed-to-parse.js received
